refactor(feature-factory): split run_factory.py into sub-modules to satisfy 400-line CI gate
run_factory.py was 1993 lines, triggering the CI file-size check. Split into:
- factory_cmd_checkpoint.py: command_checkpoint + auto-context extraction
- factory_cmd_status.py: command_status, command_repair, command_doctor
- factory_cmd_deliver.py: command_deliver, command_closeout
- factory_cmd_discover.py: command_discover
- factory_cmd_implement.py: command_implement, command_parallel, _run_serial, _run_parallel
- factory_cmd_block.py: command_block
- factory_emit.py: _emit_next_action
- factory_next_action.py: recommended_next_action
- factory_parallel.py: parse_p_annotation, parse_parallel_task_groups

Also trimmed factory_stages.py and factory_review.py which were over 400 lines.
Updated test patch targets to hit the module where each function is now defined.
All 122 tests pass.
