+++ docs/workflow/operations/codex-skills/feature-factory/scripts/run_factory.py
@@ -1,9 +1,5 @@
 #!/usr/bin/env python3
-import concurrent.futures
 import argparse
-import json
-import os
-import re
 import subprocess
 import sys
 import time
@@ -16,179 +12,88 @@ if str(_SCRIPT_DIR) not in sys.path:
 
 from factory_state import (  # noqa: E402
     REPO_ROOT,
-    FACTORY_RUNS_ROOT,
-    FACTORY_STATE,
     BLOCKED_KEY,
     DISCOVERY_KEY,
-    DELIVERY_KEY,
-    DIRTY_OVERRIDE_KEY,
-    CHECKPOINT_FALLBACK_KEY,
     CHECKPOINT_PROGRESS_KEY,
-    atomic_json_write,
-    read_json_file,
-    validated_slug,
-    normalized_repo_path,
-    workflow_dir,
-    reviews_dir,
-    scope_manifest_path,
-    factory_state_path,
+    INIT_HEAD_SHA_KEY,
     checkpoint_manifest_path,
     default_artifact_path,
     default_discovery_state,
-    default_parallel_analysis_state,
-    blocking_unresolved_items,
-    discovery_blockers_are_malformed,
-    migrate_discovery_state,
-    parse_review_frontmatter,
-    load_scope_manifest,
+    reviews_dir,
+    scope_manifest_path,
+    workflow_dir,
     save_scope_manifest,
     load_workflow_state,
     save_workflow_state,
     update_workflow_state,
-    discovery_state,
-    update_discovery_state,
     load_checkpoint_manifest,
-    PARALLEL_ANALYSIS_KEY,
-    INIT_HEAD_SHA_KEY,
+    parse_review_frontmatter,
+    blocking_unresolved_items,
+    discovery_state,
 )
 
-REVIEW_SCRIPTS = REPO_ROOT / "docs" / "workflow" / "operations" / "codex-skills" / "review-lens" / "scripts"
-if str(REVIEW_SCRIPTS) not in sys.path:
-    sys.path.insert(0, str(REVIEW_SCRIPTS))
-
-from workflow_utils import normalized_artifact_hash, repo_relative_path, resolve_stored_path  # noqa: E402
-
 from factory_git import (  # noqa: E402
-    SYNC_SCRIPT,
-    run,
-    git_output,
-    _sha_is_valid_ancestor,
     _git_head_sha,
-    current_branch_name,
-    upstream_branch_name,
-    commits_behind_upstream,
-    repo_remote_url,
+    _sha_is_valid_ancestor,
     command_path,
+    current_branch_name,
     ensure_sync,
     ensure_file,
-    create_worktree,
-    remove_all_worktrees,
-    get_new_commits,
-    stage_and_commit_if_dirty,
-    cherry_pick_commits,
-    revert_protected_files,
+    run,
+    upstream_branch_name,
 )
 
 from factory_stages import (  # noqa: E402
-    VERIFY_CHECKPOINT,
-    VERIFY_RECONCILIATION,
-    HARD_DIFF_ARTIFACT_MAX_CHARS,
-    LARGE_DIFF_RERUN_WARN_CHARS,
     CHECKPOINT_STAGES,
-    VERIFY_ON_CLOSEOUT_STAGES,
-    REQUIRED_PREDELIVERY_STAGES,
-    STAGE_PREREQUISITES,
-    STAGE_ARTIFACT_HEADINGS,
-    _CHECKPOINT_MARKER_RE,
+    VERIFY_RECONCILIATION,
+    _default_checkpoint_progress,
+    checkpoint_progress_state,
+    diff_review_budget_state,
+    later_progress_exists,
     parse_checkpoint_markers,
     parse_parallel_task_groups,
-    checkpoint_progress_state,
-    _default_checkpoint_progress,
-    artifact_has_meaningful_content,
-    verify_checkpoint_manifest,
-    diff_review_budget_state,
     preferred_diff_base_ref,
-    stage_manifest_state,
-    stage_review_inventory,
-    stage_drift_class,
-    stage_repairable,
-    stage_status_label,
-    later_progress_exists,
-    reconciliation_state,
     prerequisite_failure,
-    NEXT_ACTION_LABELS,
+    reconciliation_state,
+    stage_drift_class,
+    stage_manifest_state,
+    stage_repairable,
+    stage_review_inventory,
+    stage_status_label,
 )
 
 from factory_review import (  # noqa: E402
-    WRITE_DIFF,
-    REPAIR,
     UPDATE_REVIEW,
     APPEND_RECONCILIATION,
-    RUN_GEMINI_REVIEW,
-    RUN_CODEX_REVIEW,
-    DEFAULT_GEMINI_MODEL,
     DEFAULT_CODEX_MODEL,
     _AUTO_ACCEPT_NOTE,
-    trim_detail,
-    pick_secondary_lens,
-    required_reviews,
-    SMALL_TASK_SET_THRESHOLD,
-    resolved_review_policy,
     checkpoint_manifest,
-    repair_checkpoint_args,
-    fallback_review_command,
-    run_checkpoint_fallback,
-    record_checkpoint_fallback,
-    _advance_checkpoint_progress,
-    recommended_next_action,
+    trim_detail,
     detect_actionable_findings,
+    _advance_checkpoint_progress,
+    record_checkpoint_fallback,
+    repair_checkpoint_args,
+    required_reviews,
+    resolved_review_policy,
+    run_checkpoint_fallback,
 )
 
+from factory_emit import _emit_next_action  # noqa: E402
+from factory_next_action import recommended_next_action  # noqa: E402
+from factory_cmd_checkpoint import command_checkpoint  # noqa: E402
+from factory_cmd_discover import command_discover  # noqa: E402
 from factory_deliver import (  # noqa: E402
-    gh_json,
+    compose_closeout_text,
     current_pr_payload,
-    required_check_summary,
-    build_delivery_record,
     gather_all_review_paths,
     refresh_delivery_snapshot,
-    closeout_inventory_text,
-    compose_closeout_text,
+    required_check_summary,
 )
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-def _emit_next_action(slug: str, completed: str) -> None:
-    """Print a step-complete banner with the recommended next action.
-
-    Called after every state-advancing command so both Claude and Codex
-    orchestrators always know what to do next without running status manually.
-    Never raises — if the banner fails, the parent command already succeeded.
-    """
-    try:
-        state = load_workflow_state(slug)
-        stages = {s: stage_manifest_state(slug, s) for s in CHECKPOINT_STAGES}
-        recon_ok, _ = reconciliation_state(slug)
-        action = recommended_next_action(slug, state, stages, recon_ok)
-        label = NEXT_ACTION_LABELS.get(action, action)
-        print(f"\n[workflow] ✓ {completed}")
-        print(f"[workflow] → next: {action} — {label}")
-        if action == "mark_blocked":
-            blocked = state.get(BLOCKED_KEY, {})
-            reason = str(blocked.get("reason", "")).strip()
-            if reason:
-                print(f"[workflow]   reason: {reason}")
-        print()
-    except Exception:
-        pass  # Never fail the parent command because the banner failed
+from factory_cmd_deliver import command_deliver, command_closeout  # noqa: E402
+from factory_cmd_block import command_block  # noqa: E402
+from factory_cmd_implement import command_implement, command_parallel, _run_serial, _run_parallel  # noqa: E402
+from factory_cmd_status import command_status, command_repair, command_doctor  # noqa: E402
+from workflow_utils import resolve_stored_path  # noqa: E402
 
 
 def command_init(args: argparse.Namespace) -> int:
@@ -232,374 +137,6 @@ def command_init(args: argparse.Namespace) -> int:
     save_workflow_state(args.slug, existing_state)
     print(str(root))
     return 0
-
-
-# Module-level regex for _extract_file_paths_from_artifact — compiled once, reused per call.
-_AUTO_CONTEXT_EXTENSIONS = re.compile(
-    r"\.(tsx?|jsx?|py|prisma|sql|json|yaml|yml|md|sh|toml)$", re.IGNORECASE
-)
-_AUTO_CONTEXT_PATH_RE = re.compile(r"`([^`\n]+)`|(?<!\w)((?:cloud|docs|specs|scripts)/\S+)")
-
-# Maximum number of auto-context files injected from a single artifact. Prevents
-# large specs/plans that reference dozens of files from ballooning the review context.
-_AUTO_CONTEXT_MAX_FILES = 10
-
-
-def _extract_file_paths_from_artifact(artifact_path: Path, repo_root: Path) -> list[str]:
-    """
-    Parse a spec or plan artifact and return repo-relative paths for any files
-    explicitly listed in scope sections (Files to modify, Files to create, etc.).
-    Only includes paths that exist on disk — skips files not yet created.
-    Capped at _AUTO_CONTEXT_MAX_FILES to prevent unbounded context growth.
-    """
-    if not artifact_path.exists():
-        return []
-    text = artifact_path.read_text(encoding="utf-8")
-    found: list[str] = []
-    seen: set[str] = set()
-    for match in _AUTO_CONTEXT_PATH_RE.finditer(text):
-        if len(found) >= _AUTO_CONTEXT_MAX_FILES:
-            break
-        raw = (match.group(1) or match.group(2) or "").strip().rstrip(".,;)")
-        if not raw or "/" not in raw:
-            continue
-        if not _AUTO_CONTEXT_EXTENSIONS.search(raw):
-            continue
-        # Resolve to absolute, then back to repo-relative
-        candidate = (repo_root / raw).resolve()
-        if not candidate.exists():
-            continue
-        rel = str(candidate.relative_to(repo_root))
-        if rel not in seen:
-            seen.add(rel)
-            found.append(rel)
-    return found
-
-
-def command_checkpoint(args: argparse.Namespace) -> int:
-    fast = getattr(args, "fast", False)
-    if fast and args.stage != "diff":
-        raise SystemExit("--fast requires --stage diff")
-
-    ensure_sync()
-    root = workflow_dir(args.slug)
-    reviews = reviews_dir(args.slug)
-    root.mkdir(parents=True, exist_ok=True)
-    reviews.mkdir(parents=True, exist_ok=True)
-
-    if not fast:
-        prereq_error = prerequisite_failure(args.slug, args.stage)
-        if prereq_error:
-            raise SystemExit(prereq_error)
-    if not fast and args.stage == "tasks":
-        parallel = load_workflow_state(args.slug).get(PARALLEL_ANALYSIS_KEY, {})
-        if not parallel.get("reviewed"):
-            raise SystemExit(
-                "tasks checkpoint requires parallel analysis first; "
-                "run 'parallel --slug <slug> --note \"...\"' to record whether "
-                "parallel implementation opportunities exist in tasks.md"
-            )
-    if not fast and args.stage == "spec":
-        discovery = discovery_state(args.slug)
-        blocking = blocking_unresolved_items(discovery)
-        if blocking:
-            if discovery_blockers_are_malformed(discovery):
-                raise SystemExit(
-                    "spec checkpoint requires discovery state to be repaired first; "
-                    "use discover --clear to reset malformed discovery state"
-                )
-            raise SystemExit(
-                "spec checkpoint requires discovery unresolved items to be resolved or deferred first; "
-                "use discover --resolve or discover --defer before checkpointing spec"
-            )
-        if discovery.get("required") and not discovery.get("complete"):
-            raise SystemExit(
-                "spec checkpoint requires discovery to be complete first; "
-                "record the remaining questions and assumptions with the discover command"
-            )
-    policy = resolved_review_policy(args.slug, args)
-    context_paths = [normalized_repo_path(path, "context path") for path in args.context]
-    allow_dirty_paths = [normalized_repo_path(path, "allow-dirty path") for path in args.allow_dirty_path]
-    prior_diff_budget = diff_review_budget_state(args.slug) if args.stage == "diff" else None
-
-    # Capture HEAD SHA now — before any reviews launch — so it reflects the
-    # state of the repo at diff generation time, not after verify completes.
-    pending_head_sha: str = _git_head_sha(REPO_ROOT) or "" if args.stage == "diff" else ""
-
-    if args.stage == "diff":
-        # Determine whether a checkpoint-scoped base_ref should override the default.
-        marker_count, current_markers_sha = parse_checkpoint_markers(args.slug)
-        if marker_count == 0:
-            print(
-                "warning: tasks.md has no [CHECKPOINT] markers — diff review will cover "
-                "the full branch. Add [CHECKPOINT] markers to tasks.md to scope each "
-                "diff review to a single implementation slice.",
-                file=sys.stderr,
-            )
-        progress = checkpoint_progress_state(args.slug)
-        use_checkpoint_base = False
-        if marker_count > 0 and not args.base_ref:
-            idx = progress["index"]
-            stored_sha = progress["markers_sha"]
-            last_head = progress["last_diff_head_sha"]
-            if idx == 0:
-                pass  # first slice — use branch base as normal
-            elif idx >= marker_count:
-                print(
-                    f"warning: checkpoint_progress.index ({idx}) >= marker count ({marker_count}); "
-                    "tasks.md may have had markers removed — resetting checkpoint progress and using branch base.",
-                    file=sys.stderr,
-                )
-                update_workflow_state(
-                    args.slug,
-                    lambda s: s.__setitem__(CHECKPOINT_PROGRESS_KEY, _default_checkpoint_progress()),
-                )
-                args.base_ref = None
-            elif stored_sha != current_markers_sha:
-                print(
-                    "warning: [CHECKPOINT] marker lines changed since last checkpoint — "
-                    "resetting checkpoint progress and using branch base.",
-                    file=sys.stderr,
-                )
-                update_workflow_state(
-                    args.slug,
-                    lambda s: s.__setitem__(CHECKPOINT_PROGRESS_KEY, _default_checkpoint_progress()),
-                )
-                args.base_ref = None
-            elif _sha_is_valid_ancestor(last_head):
-                use_checkpoint_base = True
-                args.base_ref = last_head
-            else:
-                print(
-                    f"warning: last_diff_head_sha {last_head[:12] if last_head else '(empty)'} is not a "
-                    "valid ancestor of HEAD (possible rebase/amend) — "
-                    "resetting checkpoint progress and using branch base.",
-                    file=sys.stderr,
-                )
-                update_workflow_state(
-                    args.slug,
-                    lambda s: s.__setitem__(CHECKPOINT_PROGRESS_KEY, _default_checkpoint_progress()),
-                )
-                args.base_ref = None
-        _ = use_checkpoint_base  # consumed above via args.base_ref
-        args.base_ref = preferred_diff_base_ref(args.slug, args.base_ref)
-
-    scope_manifest = scope_manifest_path(args.slug)
-    if args.stage == "diff" and args.path:
-        scope_manifest = save_scope_manifest(args.slug, args.path)
-
-    artifact_path = Path(args.artifact).resolve() if args.artifact else default_artifact_path(args.slug, args.stage)
-
-    # Auto-add files mentioned in the spec/artifact as context so reviewers
-    # can verify assumptions against real code rather than generating unfounded findings.
-    if args.stage in ("spec", "plan") and not getattr(args, "no_auto_context", False):
-        auto_context = _extract_file_paths_from_artifact(artifact_path, REPO_ROOT)
-        for p in auto_context:
-            if len(context_paths) >= _AUTO_CONTEXT_MAX_FILES:
-                break
-            if p not in context_paths:
-                context_paths.append(p)
-
-    if args.stage in STAGE_ARTIFACT_HEADINGS and not artifact_has_meaningful_content(args.stage, artifact_path):
-        raise SystemExit(
-            f"{args.stage} checkpoint requires a non-stub artifact first: {artifact_path}"
-        )
-    if args.stage == "diff":
-        diff_budget = prior_diff_budget or diff_review_budget_state(args.slug)
-        if (
-            diff_budget
-            and diff_budget.get("codex_review_present")
-            and diff_budget.get("large_artifact")
-            and diff_budget.get("artifact_exists")
-            and diff_budget.get("artifact_changed_since_codex")
-            and not args.allow_large_diff_rerun
-        ):
-            raise SystemExit(
-                "Large diff rerun would regenerate the Codex review for a "
-                f"{len(artifact_path.read_text(encoding='utf-8')) if artifact_path.exists() else 0}-character artifact. "
-                "Batch more implementation fixes before rerunning the diff checkpoint, "
-                "or pass --allow-large-diff-rerun if this spend is intentional."
-            )
-    if args.stage == "diff" and not args.use_existing_artifact:
-        if not scope_manifest.exists():
-            raise SystemExit("Diff checkpoint requires a saved scope manifest or explicit --path values")
-        run(
-            [
-                sys.executable,
-                str(WRITE_DIFF),
-                "--repo",
-                str(REPO_ROOT),
-                "--output",
-                str(artifact_path),
-                "--path-manifest",
-                str(scope_manifest),
-                *sum((["--allow-dirty-path", path] for path in allow_dirty_paths), []),
-                *([] if not args.base_ref else ["--base-ref", args.base_ref]),
-            ]
-        )
-        diff_meta_path = artifact_path.with_suffix(artifact_path.suffix + ".json")
-        if diff_meta_path.exists():
-            diff_meta = json.loads(diff_meta_path.read_text(encoding="utf-8"))
-            args.base_ref = diff_meta.get("git_base_ref") or args.base_ref
-    elif not artifact_path.exists():
-        raise SystemExit(f"Artifact does not exist: {artifact_path}")
-
-    if args.stage == "diff":
-        diff_meta_path = artifact_path.with_suffix(artifact_path.suffix + ".json")
-        if diff_meta_path.exists():
-            diff_meta = json.loads(diff_meta_path.read_text(encoding="utf-8"))
-            args.base_ref = diff_meta.get("git_base_ref") or args.base_ref
-            if args.use_existing_artifact:
-                recorded_head = diff_meta.get("git_head_sha", "")
-                if recorded_head:
-                    current_head = _git_head_sha(REPO_ROOT)
-                    if current_head and current_head != recorded_head:
-                        raise SystemExit(
-                            f"Existing diff artifact is stale: recorded HEAD {recorded_head[:12]} "
-                            f"does not match current HEAD {current_head[:12]}. "
-                            "Drop --use-existing-artifact to regenerate the diff, "
-                            "or pass --base-ref if you intentionally want to review an older slice."
-                        )
-
-    if args.stage == "diff" and artifact_path.exists():
-        diff_text = artifact_path.read_text(encoding="utf-8")
-        if len(diff_text) > HARD_DIFF_ARTIFACT_MAX_CHARS:
-            raise SystemExit(
-                f"Diff artifact exceeds hard cap ({len(diff_text)} > {HARD_DIFF_ARTIFACT_MAX_CHARS}). "
-                "Split the review scope into smaller workflow paths or use a smaller diff artifact."
-            )
-        if len(diff_text) >= LARGE_DIFF_RERUN_WARN_CHARS:
-            print(
-                "warning: large diff artifact detected; future diff reruns will regenerate "
-                "the Codex review unless the artifact stays unchanged. Batch follow-up fixes when possible.",
-                file=sys.stderr,
-            )
-
-    reviews = getattr(args, "required_reviews", None)
-    if reviews is None:
-        small_task_set = False
-        if args.stage in ("tasks", "closeout"):
-            # For "tasks": count from the artifact (tasks.md is the artifact).
-            # For "closeout": count from tasks.md in the workflow dir (artifact is closeout.md).
-            tasks_file = artifact_path if args.stage == "tasks" else workflow_dir(args.slug) / "tasks.md"
-            if tasks_file.exists():
-                try:
-                    task_text = tasks_file.read_text(encoding="utf-8")
-                    task_count = sum(
-                        1 for line in task_text.splitlines()
-                        if line.strip().startswith("- [ ]") or line.strip().startswith("- [x]") or line.strip().startswith("- [X]")
-                    )
-                    if task_count < SMALL_TASK_SET_THRESHOLD:
-                        small_task_set = True
-                        print(
-                            f"info: small task set ({task_count} tasks < {SMALL_TASK_SET_THRESHOLD} threshold) — "
-                            f"skipping Gemini reviews for {args.stage} stage, running Codex review only.",
-                            file=sys.stderr,
-                        )
-                except Exception:
-                    pass  # count failed — fall back to full reviews
-        reviews = required_reviews(
-            args.stage,
-            policy["sensitive"],
-            policy["large_structural"],
-            policy["performance_sensitive"],
-            policy["extra_gemini_lenses"],
-            fast=fast,
-            small_task_set=small_task_set,
-        )
-
-    manifest = checkpoint_manifest(
-        args.slug,
-        args.stage,
-        artifact_path,
-        args.base_ref,
-        context_paths,
-        reviews,
-        args.max_artifact_chars,
-        args.max_context_chars,
-        args.max_total_chars,
-        allow_dirty_paths,
-    )
-    manifest_path = checkpoint_manifest_path(args.slug, args.stage)
-    atomic_json_write(manifest_path, manifest)
-    if args.stage == "diff":
-        if allow_dirty_paths:
-            update_workflow_state(
-                args.slug,
-                lambda state: state.__setitem__(
-                    DIRTY_OVERRIDE_KEY,
-                    {
-                        "allowed_dirty_paths": allow_dirty_paths,
-                        "used": True,
-                        "updated_at": int(time.time()),
-                    },
-                ),
-            )
-        else:
-            # No --allow-dirty-path on this run — clear any stale override from a prior run.
-            update_workflow_state(
-                args.slug,
-                lambda state: state.pop(DIRTY_OVERRIDE_KEY, None),
-            )
-
-    cmd = [
-        sys.executable,
-        str(REPAIR),
-        "--checkpoint-manifest",
-        str(manifest_path),
-        "--workspace-dir",
-        str(REPO_ROOT),
-        "--gemini-timeout-seconds",
-        str(args.gemini_timeout_seconds),
-        "--gemini-retries",
-        str(args.gemini_retries),
-    ]
-    try:
-        result = subprocess.run(
-            cmd,
-            text=True,
-            timeout=args.repair_timeout_seconds,
-        )
-    except subprocess.TimeoutExpired:
-        if not args.fallback:
-            print(
-                f"checkpoint blocked: repair exceeded {args.repair_timeout_seconds}s for "
-                f"{args.stage} on {args.slug}",
-                file=sys.stderr,
-            )
-            return 1
-        fallback_reason = f"repair exceeded {args.repair_timeout_seconds}s"
-    else:
-        if result.returncode == 0:
-            reverted = revert_protected_files()
-            if reverted:
-                print(f"reverted protected files: {', '.join(reverted)}", file=sys.stderr)
-            _advance_checkpoint_progress(args.slug, args.stage, pending_head_sha)
-            _emit_next_action(args.slug, f"{args.stage} checkpoint")
-            return 0
-        if not args.fallback:
-            return result.returncode
-        fallback_reason = f"repair exited {result.returncode}"
-
-    fallback_ok, fallback_detail = run_checkpoint_fallback(
-        manifest_path,
-        REPO_ROOT,
-        args.gemini_timeout_seconds,
-        args.gemini_retries,
-    )
-    if not fallback_ok:
-        print(f"checkpoint blocked: fallback review path failed for {args.stage} on {args.slug}: {trim_detail(fallback_detail)}", file=sys.stderr)
-        return 1
-    reverted = revert_protected_files()
-    if reverted:
-        print(f"reverted protected files: {', '.join(reverted)}", file=sys.stderr)
-    record_checkpoint_fallback(args.slug, args.stage, fallback_reason)
-    _advance_checkpoint_progress(args.slug, args.stage, pending_head_sha)
-    print(f"warning: fallback checkpoint path used for {args.stage} on {args.slug}: {fallback_reason}", file=sys.stderr)
-    _emit_next_action(args.slug, f"{args.stage} checkpoint")
-    return 0
-
-
 
 
 def command_reconcile(args: argparse.Namespace) -> int:
@@ -682,1158 +219,6 @@ def command_auto_reconcile(args: argparse.Namespace) -> int:
     if not auto_accepted and not needs_review:
         print(f"all reviews for {args.stage} already reconciled")
     _emit_next_action(args.slug, f"auto-reconcile {args.stage}")
-    return 0
-
-
-def command_block(args: argparse.Namespace) -> int:
-    ensure_sync()
-    if not args.clear and not args.reason:
-        raise SystemExit("block requires --reason unless --clear is used")
-
-    def mutate(state: dict) -> None:
-        if args.clear:
-            state[BLOCKED_KEY] = {
-                "active": False,
-                "reason": "",
-                "updated_at": int(time.time()),
-            }
-            return
-        state[BLOCKED_KEY] = {
-            "active": True,
-            "reason": args.reason,
-            "updated_at": int(time.time()),
-        }
-
-    state = update_workflow_state(args.slug, mutate)
-    blocked = state.get(BLOCKED_KEY, {})
-    if blocked.get("active"):
-        print(f"blocked: {blocked.get('reason', '')}")
-    else:
-        print("unblocked")
-    return 0
-
-
-def command_discover(args: argparse.Namespace) -> int:
-    ensure_sync()
-    clear = getattr(args, "clear", False)
-    force_complete = getattr(args, "force_complete", False)
-    if clear and any(
-        [
-            args.required,
-            args.count is not None,
-            args.question,
-            args.recommendation,
-            args.rationale,
-            args.assumption,
-            args.summary is not None,
-            args.complete,
-            force_complete,
-            getattr(args, "unresolved", None) is not None,
-            getattr(args, "resolve", None) is not None,
-            getattr(args, "defer", None) is not None,
-            getattr(args, "non_goal", None) is not None,
-            getattr(args, "acceptance_criteria", None) is not None,
-            getattr(args, "answer", None) is not None,
-        ]
-    ):
-        raise SystemExit("discover --clear cannot be combined with other discovery updates")
-    if args.count is not None and args.count < 0:
-        raise SystemExit("discover requires --count to be zero or greater")
-    if any([args.question, args.recommendation, args.rationale]) and not all(
-        [args.question, args.recommendation, args.rationale]
-    ):
-        raise SystemExit("discover requires --question, --recommendation, and --rationale together")
-    if not clear and not any(
-        [
-            args.required,
-            args.count is not None,
-            args.question,
-            args.assumption,
-            args.summary is not None,
-            args.complete,
-            force_complete,
-            getattr(args, "unresolved", None) is not None,
-            getattr(args, "resolve", None) is not None,
-            getattr(args, "defer", None) is not None,
-            getattr(args, "non_goal", None) is not None,
-            getattr(args, "acceptance_criteria", None) is not None,
-            getattr(args, "answer", None) is not None,
-        ]
-    ):
-        raise SystemExit("discover requires at least one update, or use --clear to reset discovery state")
-
-    def mutate(discovery: dict) -> None:
-        if clear:
-            preserved_required = bool(discovery.get("required"))
-            preserved_question_count = discovery.get("question_count", 0)
-            if not isinstance(preserved_question_count, int):
-                preserved_question_count = 0
-            preserved_asked_count = discovery.get("asked_count", 0)
-            if not isinstance(preserved_asked_count, int):
-                preserved_asked_count = 0
-            preserved_questions = discovery.get("questions", [])
-            if not isinstance(preserved_questions, list):
-                preserved_questions = []
-            preserved_assumptions = discovery.get("assumptions", [])
-            if not isinstance(preserved_assumptions, list):
-                preserved_assumptions = []
-            preserved_summary = discovery.get("summary", "")
-            if not isinstance(preserved_summary, str):
-                preserved_summary = ""
-            preserved_answers = discovery.get("answers", {})
-            if not isinstance(preserved_answers, dict):
-                preserved_answers = {}
-            preserved_non_goals = discovery.get("non_goals", [])
-            if not isinstance(preserved_non_goals, list):
-                preserved_non_goals = []
-            preserved_acceptance = discovery.get("acceptance_criteria", [])
-            if not isinstance(preserved_acceptance, list):
-                preserved_acceptance = []
-            discovery.clear()
-            discovery.update(default_discovery_state())
-            discovery["required"] = (
-                preserved_required
-                or bool(preserved_question_count)
-                or bool(preserved_questions)
-                or bool(preserved_assumptions)
-            )
-            discovery["question_count"] = preserved_question_count
-            discovery["asked_count"] = preserved_asked_count
-            discovery["questions"] = preserved_questions
-            discovery["assumptions"] = preserved_assumptions
-            discovery["summary"] = preserved_summary
-            discovery["answers"] = preserved_answers
-            discovery["non_goals"] = preserved_non_goals
-            discovery["acceptance_criteria"] = preserved_acceptance
-            discovery["unresolved"] = []
-            discovery["complete"] = False
-            discovery["updated_at"] = int(time.time())
-            return
-        if args.required:
-            discovery["required"] = True
-        if args.count is not None:
-            if discovery.get("asked_count", 0) > args.count:
-                raise SystemExit(
-                    "discover requires --count to stay at or above the number of already asked questions; "
-                    "use --clear to restart discovery"
-                )
-            discovery["question_count"] = args.count
-            discovery["required"] = discovery["required"] or args.count > 0
-        if args.question:
-            questions = list(discovery.get("questions", []))
-            questions.append(
-                {
-                    "question": args.question,
-                    "recommendation": args.recommendation,
-                    "rationale": args.rationale,
-                    "updated_at": int(time.time()),
-                }
-            )
-            discovery["questions"] = questions
-            discovery["asked_count"] = len(questions)
-            discovery["required"] = True
-            if discovery.get("question_count", 0) < len(questions):
-                discovery["question_count"] = len(questions)
-        if args.assumption:
-            assumptions = list(discovery.get("assumptions", []))
-            for assumption in args.assumption:
-                if assumption not in assumptions:
-                    assumptions.append(assumption)
-            discovery["assumptions"] = assumptions
-            discovery["required"] = discovery["required"] or bool(assumptions)
-        if args.summary is not None:
-            discovery["summary"] = args.summary
-            discovery["required"] = discovery["required"] or bool(args.summary.strip())
-        if getattr(args, "answer", None) is not None:
-            question_text, answer_text = args.answer
-            if not isinstance(discovery.get("answers"), dict):
-                discovery["answers"] = {}
-            discovery["answers"][question_text] = answer_text
-        if getattr(args, "unresolved", None) is not None:
-            item_text = args.unresolved
-            unresolved = discovery.setdefault("unresolved", [])
-            if not any(u["item"] == item_text for u in unresolved):
-                unresolved.append({"item": item_text, "deferred": False})
-        if getattr(args, "resolve", None) is not None:
-            resolve_text = args.resolve
-            discovery["unresolved"] = [
-                u for u in discovery.get("unresolved", []) if u["item"] != resolve_text
-            ]
-        if getattr(args, "defer", None) is not None:
-            defer_text = args.defer
-            for u in discovery.get("unresolved", []):
-                if u["item"] == defer_text:
-                    u["deferred"] = True
-                    break
-        if getattr(args, "non_goal", None) is not None:
-            ng = discovery.setdefault("non_goals", [])
-            if args.non_goal not in ng:
-                ng.append(args.non_goal)
-        if getattr(args, "acceptance_criteria", None) is not None:
-            ac = discovery.setdefault("acceptance_criteria", [])
-            if args.acceptance_criteria not in ac:
-                ac.append(args.acceptance_criteria)
-        blocking = blocking_unresolved_items(discovery)
-        if args.complete:
-            if blocking:
-                if discovery_blockers_are_malformed(discovery):
-                    raise SystemExit(
-                        "discover cannot mark discovery complete while discovery state is malformed; "
-                        "use discover --clear to reset malformed discovery state"
-                    )
-                raise SystemExit(
-                    "discover cannot mark discovery complete while unresolved items remain; "
-                    "resolve or defer each item first"
-                )
-            if not force_complete and int(discovery.get("asked_count", 0)) < int(discovery.get("question_count", 0)):
-                raise SystemExit(
-                    "discover cannot mark discovery complete before the planned questions are recorded; "
-                    "use --force-complete if you intentionally want to override the count"
-                )
-            discovery["complete"] = True
-        elif (
-            args.required
-            or args.count is not None
-            or args.question
-            or args.assumption
-            or args.summary is not None
-            or getattr(args, "answer", None) is not None
-            or getattr(args, "unresolved", None) is not None
-            or getattr(args, "resolve", None) is not None
-            or getattr(args, "defer", None) is not None
-            or getattr(args, "non_goal", None) is not None
-            or getattr(args, "acceptance_criteria", None) is not None
-        ):
-            discovery["complete"] = False
-        if force_complete:
-            if blocking:
-                raise SystemExit(
-                    "discover cannot force discovery complete while unresolved items remain; "
-                    "resolve or defer each item first"
-                )
-            discovery["complete"] = True
-        discovery["updated_at"] = int(time.time())
-
-    state = update_discovery_state(args.slug, mutate)
-    discovery = state.get(DISCOVERY_KEY, {})
-    remaining = max(int(discovery.get("question_count", 0)) - int(discovery.get("asked_count", 0)), 0)
-    print(f"workflow: {args.slug}")
-    print("discovery:")
-    print(f"- version: {discovery.get('version', 1)}")
-    print(f"- required: {'yes' if discovery.get('required') else 'no'}")
-    print(f"- complete: {'yes' if discovery.get('complete') else 'no'}")
-    print(f"- question-count: {discovery.get('question_count', 0)}")
-    print(f"- asked-count: {discovery.get('asked_count', 0)}")
-    print(f"- remaining: {remaining}")
-    if discovery.get("assumptions"):
-        print(f"- assumptions: {len(discovery.get('assumptions', []))}")
-        for assumption in discovery.get("assumptions", []):
-            print(f"- assumption: {assumption}")
-    if discovery.get("summary"):
-        print(f"- summary: {trim_detail(str(discovery.get('summary', '')))}")
-    if discovery.get("non_goals"):
-        print(f"- non-goals: {len(discovery['non_goals'])}")
-        for ng in discovery["non_goals"]:
-            print(f"  - {ng}")
-    if discovery.get("acceptance_criteria"):
-        print(f"- acceptance-criteria: {len(discovery['acceptance_criteria'])}")
-        for ac in discovery["acceptance_criteria"]:
-            print(f"  - {ac}")
-    if discovery.get("answers"):
-        print(f"- answers: {len(discovery['answers'])}")
-    if discovery.get("unresolved"):
-        print(f"- unresolved: {len(discovery['unresolved'])}")
-        for u in discovery["unresolved"]:
-            status = " [deferred]" if u.get("deferred") else ""
-            print(f"  - {u['item']}{status}")
-    if args.complete or getattr(args, "force_complete", False):
-        _emit_next_action(args.slug, "discovery complete")
-    return 0
-
-
-
-
-def command_status(args: argparse.Namespace) -> int:
-    ensure_sync()
-    state = load_workflow_state(args.slug)
-    stages = {stage: stage_manifest_state(args.slug, stage) for stage in CHECKPOINT_STAGES}
-    recon_ok, recon_detail = reconciliation_state(args.slug)
-    next_action = recommended_next_action(args.slug, state, stages, recon_ok)
-    branch = current_branch_name() or "(detached HEAD)"
-    upstream = upstream_branch_name() or "(none)"
-    blocked = state.get(BLOCKED_KEY, {})
-    delivery = refresh_delivery_snapshot(state.get(DELIVERY_KEY, {}))
-    discovery = discovery_state(args.slug)
-    dirty_override = state.get(DIRTY_OVERRIDE_KEY, {})
-    checkpoint_fallback = state.get(CHECKPOINT_FALLBACK_KEY, {})
-    diff_budget = diff_review_budget_state(args.slug)
-
-    behind = commits_behind_upstream()
-
-    print(f"workflow: {args.slug}")
-    print(f"branch: {branch}")
-    print(f"upstream: {upstream}")
-    if behind is not None and behind > 0:
-        print(f"warning: branch is {behind} commit{'s' if behind != 1 else ''} behind upstream — rebase before creating PR")
-    print("")
-    print("stages:")
-    for stage in CHECKPOINT_STAGES:
-        stage_state = stages[stage]
-        status = stage_status_label(args.slug, stage, stage_state)
-        detail = trim_detail(str(stage_state.get("detail", "")))
-        _, orphaned = stage_review_inventory(args.slug, stage)
-        if orphaned:
-            orphaned_note = f"orphaned historical reviews: {len(orphaned)}"
-            detail = f"{detail}; {orphaned_note}" if detail else orphaned_note
-        print(f"- {stage}: {status}" + (f" ({detail})" if detail else ""))
-
-    print("")
-    print("reconciliation:")
-    if recon_ok:
-        print("- terminal")
-    else:
-        print(f"- blocked ({trim_detail(recon_detail)})")
-
-    print("")
-    print("blocked-state:")
-    if blocked.get("active"):
-        print(f"- active: {blocked.get('reason', '')}")
-    else:
-        print("- inactive")
-
-    blocking = blocking_unresolved_items(discovery)
-    unresolved = discovery.get("unresolved", [])
-    if discovery.get("required") or not discovery.get("complete") or discovery.get("asked_count") or discovery.get("assumptions") or discovery.get("summary") or blocking or (isinstance(unresolved, list) and unresolved):
-        print("")
-        print("discovery:")
-        print(f"- required: {'yes' if discovery.get('required') else 'no'}")
-        print(f"- complete: {'yes' if discovery.get('complete') else 'no'}")
-        print(f"- question-count: {discovery.get('question_count', 0)}")
-        print(f"- asked-count: {discovery.get('asked_count', 0)}")
-        remaining = max(int(discovery.get("question_count", 0)) - int(discovery.get("asked_count", 0)), 0)
-        print(f"- remaining: {remaining}")
-        if discovery.get("assumptions"):
-            print(f"- assumptions: {len(discovery.get('assumptions', []))}")
-        if discovery.get("summary"):
-            print(f"- summary: {trim_detail(str(discovery.get('summary', '')))}")
-        if isinstance(unresolved, list) and unresolved:
-            print(f"- unresolved-open: {len(blocking)}")
-            print(f"- unresolved-deferred: {len(unresolved) - len(blocking)}")
-            if blocking:
-                if discovery_blockers_are_malformed(discovery):
-                    print("- action: use discover --clear to repair malformed discovery state")
-                else:
-                    print("- action: resolve or defer unresolved items before spec")
-        elif blocking:
-            print(f"- unresolved-open: {len(blocking)}")
-            if discovery_blockers_are_malformed(discovery):
-                print("- action: use discover --clear to repair malformed discovery state")
-            else:
-                print("- action: resolve or defer unresolved items before spec")
-
-    cp = checkpoint_progress_state(args.slug)
-    if cp["index"] > 0:
-        print("")
-        print("checkpoint-progress:")
-        print(f"- index: {cp['index']}")
-        last = cp["last_diff_head_sha"]
-        print(f"- last-diff-head: {last[:12] if last else '(none)'}")
-
-    print("")
-    print("delivery:")
-    if delivery.get("pr_url"):
-        line = f"- pr: {delivery.get('pr_url')} [{delivery.get('pr_state', '')}]"
-        if delivery.get("pr_is_draft"):
-            line += " draft"
-        print(line)
-        print(f"- checks: {delivery.get('checks_summary', 'unknown')}")
-        if delivery.get("merge_state_status"):
-            print(f"- merge-state: {delivery.get('merge_state_status')}")
-        if delivery.get("head_mismatch"):
-            print("- head-mismatch: local HEAD differs from PR head")
-            print("- action: refresh the PR head before delivering")
-    else:
-        print("- not-started")
-
-    if dirty_override.get("used"):
-        print("")
-        print("dirty-override:")
-        print("- used: yes")
-        for path in dirty_override.get("allowed_dirty_paths", []):
-            print(f"- allow-dirty-path: {path}")
-
-    if checkpoint_fallback.get("used"):
-        print("")
-        print("checkpoint-fallback:")
-        print(f"- stage: {checkpoint_fallback.get('stage', '')}")
-        print(f"- reason: {checkpoint_fallback.get('reason', '')}")
-
-    if diff_budget.get("artifact_exists"):
-        print("")
-        print("diff-review-budget:")
-        print(f"- artifact-bytes: {diff_budget.get('artifact_bytes', 0)}")
-        if diff_budget.get("recorded_base_ref"):
-            base_ref = str(diff_budget.get("recorded_base_ref", ""))
-            base_sha = str(diff_budget.get("recorded_base_sha", ""))[:12]
-            print(f"- recorded-base: {base_ref} [{base_sha}]")
-        if diff_budget.get("large_artifact"):
-            print(f"- large-artifact: yes (>= {LARGE_DIFF_RERUN_WARN_CHARS} bytes)")
-        else:
-            print("- large-artifact: no")
-        if diff_budget.get("head_mismatch"):
-            recorded_head = str(diff_budget.get("recorded_head_sha", ""))[:12]
-            current_head = str(diff_budget.get("current_head_sha", ""))[:12]
-            print(f"- rerun-likely: yes (diff artifact HEAD {recorded_head} != current HEAD {current_head})")
-            print(
-                f"- scope-basis: last-reviewed-head [{recorded_head}]"
-            )
-        elif diff_budget.get("artifact_changed_since_codex"):
-            print("- rerun-likely: yes (artifact changed since the last Codex review)")
-        elif diff_budget.get("suggested_base_ref"):
-            suggested = str(diff_budget.get("suggested_base_ref", ""))[:12]
-            print(f"- scope-basis: {diff_budget.get('scope_basis', 'branch-merge-base')} [{suggested}]")
-        else:
-            print("- rerun-likely: no")
-        if diff_budget.get("large_artifact") and (
-            diff_budget.get("head_mismatch") or diff_budget.get("artifact_changed_since_codex")
-        ):
-            print("- advice: batch more implementation fixes before rerunning the diff checkpoint")
-
-    print("")
-    print(f"next-action: {next_action}")
-    return 0
-
-
-def command_repair(args: argparse.Namespace) -> int:
-    ensure_sync()
-    state = load_workflow_state(args.slug)
-    stages = {stage: stage_manifest_state(args.slug, stage) for stage in CHECKPOINT_STAGES}
-    repaired: list[str] = []
-    notes: list[str] = []
-    blocked_reason = ""
-
-    print(f"workflow: {args.slug}")
-    print("repair:")
-
-    for stage in ["spec", "plan", "tasks", "diff"]:
-        stage_state = stages[stage]
-        drift = stage_drift_class(stage, stage_state)
-        _, orphaned = stage_review_inventory(args.slug, stage)
-        if stage_repairable(args.slug, stage, stage_state):
-            print(f"- {stage}: repairing {drift}")
-            result = command_checkpoint(repair_checkpoint_args(args.slug, stage, stage_state))
-            if result != 0:
-                blocked_reason = f"{stage} repair failed"
-                break
-            refreshed = stage_manifest_state(args.slug, stage)
-            stages[stage] = refreshed
-            if not refreshed["healthy"]:
-                blocked_reason = f"{stage} remains unhealthy: {trim_detail(str(refreshed.get('detail', '')))}"
-                break
-            repaired.append(stage)
-            if orphaned:
-                notes.append(f"{stage}: preserved {len(orphaned)} orphaned historical review file(s)")
-            continue
-
-        print(f"- {stage}: {stage_status_label(args.slug, stage, stage_state)}")
-
-        if drift in {"missing-artifact", "stub-artifact"} and later_progress_exists(stages, stage)[0]:
-            blocked_reason = f"{stage} has {drift.replace('-', ' ')} while later workflow material exists"
-            break
-        if drift == "unhealthy-manifest" and not stage_repairable(args.slug, stage, stage_state):
-            blocked_reason = f"{stage} checkpoint is not safely repairable: {trim_detail(str(stage_state.get('detail', '')))}"
-            break
-
-    if not blocked_reason:
-        # Repair closeout if manifest is stale (only when it exists and is unhealthy)
-        closeout_state = stages["closeout"]
-        closeout_drift = stage_drift_class("closeout", closeout_state)
-        if closeout_drift == "unhealthy-manifest" and stage_repairable(args.slug, "closeout", closeout_state):
-            print("- closeout: repairing unhealthy-manifest")
-            result = command_checkpoint(repair_checkpoint_args(args.slug, "closeout", closeout_state))
-            if result != 0:
-                blocked_reason = "closeout repair failed"
-            else:
-                refreshed = stage_manifest_state(args.slug, "closeout")
-                stages["closeout"] = refreshed
-                if not refreshed["healthy"]:
-                    blocked_reason = f"closeout remains unhealthy: {trim_detail(str(refreshed.get('detail', '')))}"
-                else:
-                    repaired.append("closeout")
-        elif closeout_drift == "unhealthy-manifest":
-            # unhealthy-manifest but not repairable — block so repair doesn't silently succeed
-            blocked_reason = "closeout is unhealthy but not repairable"
-        elif closeout_drift not in {"not-checkpointed", "missing-artifact", "stub-artifact"}:
-            print(f"- closeout: {stage_status_label(args.slug, 'closeout', closeout_state)}")
-
-    if blocked_reason:
-        print(f"- blocked: {blocked_reason}")
-        for note in notes:
-            print(f"- note: {note}")
-        return 1
-
-    recon_ok, recon_detail = reconciliation_state(args.slug)
-    next_action = recommended_next_action(args.slug, state, stages, recon_ok)
-    print("")
-    if repaired:
-        print(f"- repaired: {', '.join(repaired)}")
-    else:
-        print("- repaired: none")
-    if notes:
-        for note in notes:
-            print(f"- note: {note}")
-    if not recon_ok:
-        print(f"- follow-up: reconciliation still blocked ({trim_detail(recon_detail)})")
-    print(f"next-action: {next_action}")
-    return 0
-
-
-def command_doctor(args: argparse.Namespace) -> int:
-    checks: list[tuple[str, str, str]] = []
-
-    def add(name: str, level: str, detail: str) -> None:
-        checks.append((name, level, detail))
-
-    repo_root = git_output("rev-parse", "--show-toplevel")
-    add("repo-root", "ok" if repo_root == str(REPO_ROOT) else "fail", repo_root or "not inside a git repo")
-    add(
-        "python",
-        "ok" if sys.version_info >= (3, 10) else "fail",
-        f"{sys.version_info.major}.{sys.version_info.minor}.{sys.version_info.micro}",
-    )
-
-    for label, path in {
-        "sync-script": SYNC_SCRIPT,
-        "feature-runner": Path(__file__).resolve(),
-        "write-diff": WRITE_DIFF,
-        "repair": REPAIR,
-        "verify-checkpoint": VERIFY_CHECKPOINT,
-        "verify-reconciliation": VERIFY_RECONCILIATION,
-    }.items():
-        add(label, "ok" if path.exists() else "fail", str(path))
-
-    for tool_name in ["git", "codex", "gemini", "gh"]:
-        found = command_path(tool_name)
-        level = "ok" if found else ("warn" if tool_name == "gh" else "fail")
-        add(f"tool:{tool_name}", level, found or "not installed")
-
-    sync_check = subprocess.run([sys.executable, str(SYNC_SCRIPT), "--check"], text=True, capture_output=True)
-    add("skill-sync", "ok" if sync_check.returncode == 0 else "warn", "in sync" if sync_check.returncode == 0 else "needs sync")
-
-    branch = current_branch_name()
-    add("git-branch", "ok" if branch else "warn", branch or "detached HEAD")
-    upstream = upstream_branch_name()
-    add("git-upstream", "ok" if upstream else "warn", upstream or "no upstream configured")
-
-    if command_path("gh"):
-        auth = subprocess.run(["gh", "auth", "status", "--active"], text=True, capture_output=True)
-        auth_detail = (auth.stderr or auth.stdout or "").strip().splitlines()
-        add("gh-auth", "ok" if auth.returncode == 0 else "warn", auth_detail[0] if auth_detail else "unknown")
-
-    failed = False
-    for name, level, detail in checks:
-        print(f"[{level.upper()}] {name}: {detail}")
-        if level == "fail":
-            failed = True
-    return 1 if failed else 0
-
-
-def command_deliver(args: argparse.Namespace) -> int:
-    ensure_sync()
-    if not command_path("gh"):
-        raise SystemExit("deliver requires the gh CLI to be installed")
-
-    for stage in REQUIRED_PREDELIVERY_STAGES:
-        manifest_path = checkpoint_manifest_path(args.slug, stage)
-        if not manifest_path.exists():
-            raise SystemExit(f"deliver requires completed {stage} checkpoint first")
-        healthy, detail = verify_checkpoint_manifest(manifest_path)
-        if not healthy:
-            raise SystemExit(f"deliver requires a healthy {stage} checkpoint first: {trim_detail(detail)}")
-
-    recon_ok, recon_detail = reconciliation_state(args.slug)
-    if not recon_ok:
-        raise SystemExit(f"deliver requires terminal reconciliation first: {trim_detail(recon_detail)}")
-
-    auth = subprocess.run(["gh", "auth", "status", "--active"], text=True, capture_output=True)
-    if auth.returncode != 0:
-        raise SystemExit(trim_detail(auth.stderr or auth.stdout or "GitHub authentication is not ready"))
-
-    branch = current_branch_name()
-    if not branch:
-        raise SystemExit("deliver requires a named branch; detached HEAD is not supported")
-    upstream = upstream_branch_name()
-    head_sha = git_output("rev-parse", "HEAD") or ""
-    diff_budget = diff_review_budget_state(args.slug)
-    if diff_budget.get("head_mismatch"):
-        recorded_head = str(diff_budget.get("recorded_head_sha", ""))[:12]
-        current_head = str(diff_budget.get("current_head_sha", ""))[:12]
-        raise SystemExit(
-            f"deliver requires the current branch to match the reviewed diff HEAD; "
-            f"diff artifact HEAD {recorded_head} does not match current HEAD {current_head}. "
-            "Rerun the diff checkpoint before delivering."
-        )
-
-    pr = current_pr_payload()
-    if pr and pr.get("state") != "OPEN":
-        pr = None
-
-    if args.create_pr and not pr:
-        if not upstream:
-            raise SystemExit("deliver requires a published branch with an upstream before creating a PR")
-        behind = commits_behind_upstream()
-        if behind is not None and behind > 0:
-            print(f"warning: branch is {behind} commit{'s' if behind != 1 else ''} behind upstream — consider rebasing before creating PR")
-        title = args.title or f"Workflow: {args.slug}"
-        body = "\n".join(
-            [
-                f"## Workflow",
-                f"- slug: `{args.slug}`",
-                f"- spec: `docs/workflow/feature-runs/{args.slug}/spec.md`",
-                f"- plan: `docs/workflow/feature-runs/{args.slug}/plan.md`",
-                f"- tasks: `docs/workflow/feature-runs/{args.slug}/tasks.md`",
-                "",
-                "## Verification",
-                "- Local workflow checkpoints completed through diff review.",
-            ]
-        )
-        cmd = [
-            "gh",
-            "pr",
-            "create",
-            "--head",
-            branch,
-            "--title",
-            title,
-            "--body",
-            body,
-        ]
-        if args.base:
-            cmd.extend(["--base", args.base])
-        if args.draft:
-            cmd.append("--draft")
-        if args.dry_run:
-            cmd.append("--dry-run")
-        result = subprocess.run(cmd, text=True, capture_output=True)
-        if result.returncode != 0:
-            raise SystemExit(trim_detail(result.stderr or result.stdout or "PR creation failed"))
-        if args.dry_run:
-            print((result.stdout or "").strip())
-        else:
-            pr = current_pr_payload()
-
-    if (args.watch_ci or args.merge_when_green or args.auto_merge) and not pr:
-        raise SystemExit("deliver requires an open PR before watching checks or merging")
-
-    checks_summary = "unknown"
-    checks: list[dict] = []
-    checks_detail = ""
-    if pr:
-        head_mismatch = bool(pr.get("headRefOid")) and bool(head_sha) and pr.get("headRefOid") != head_sha
-        if head_mismatch and (args.merge_when_green or args.auto_merge):
-            raise SystemExit(
-                f"local HEAD {head_sha[:12]} does not match PR head {str(pr.get('headRefOid'))[:12]}; "
-                "push the branch or refresh the PR before merging"
-            )
-        checks_summary, checks, checks_detail = required_check_summary(
-            pr.get("number"),
-            watch=args.watch_ci and not args.dry_run,
-            interval=args.interval,
-        )
-        delivery_record = build_delivery_record(pr, checks_summary, checks, branch, head_sha)
-        delivery_record["upstream"] = upstream or ""
-        delivery_record["checks_detail"] = checks_detail
-        delivery_record["head_mismatch"] = head_mismatch
-
-        if args.dry_run:
-            print(f"branch: {branch}")
-            print(f"head: {head_sha}")
-            print(f"pr: {delivery_record.get('pr_url', '') or 'not created'}")
-            print(f"checks: {checks_summary}")
-            if delivery_record.get("merge_state_status"):
-                print(f"merge-state: {delivery_record.get('merge_state_status')}")
-            if head_mismatch:
-                print("head-mismatch: yes")
-            return 0
-
-        if args.auto_merge and not args.dry_run:
-            merge_cmd = [
-                "gh",
-                "pr",
-                "merge",
-                str(pr["number"]),
-                "--auto",
-                "--squash",
-                "--match-head-commit",
-                head_sha,
-            ]
-            result = subprocess.run(merge_cmd, text=True, capture_output=True)
-            if result.returncode != 0:
-                raise SystemExit(trim_detail(result.stderr or result.stdout or "Auto-merge enable failed"))
-            refreshed_delivery = refresh_delivery_snapshot(delivery_record)
-            update_workflow_state(
-                args.slug,
-                lambda state: state.__setitem__(DELIVERY_KEY, refreshed_delivery),
-            )
-        elif args.merge_when_green and not args.dry_run:
-            if checks_summary != "pass":
-                raise SystemExit(f"cannot merge while required checks are {checks_summary}")
-            merged_pr_number = pr["number"]
-            merge_cmd = [
-                "gh",
-                "pr",
-                "merge",
-                str(merged_pr_number),
-                "--squash",
-                "--match-head-commit",
-                head_sha,
-            ]
-            result = subprocess.run(merge_cmd, text=True, capture_output=True)
-            if result.returncode != 0:
-                raise SystemExit(trim_detail(result.stderr or result.stdout or "Merge failed"))
-            pr = current_pr_payload()
-            checks_summary, checks, checks_detail = required_check_summary(merged_pr_number, watch=False)
-            update_workflow_state(
-                args.slug,
-                lambda state: state.__setitem__(
-                    DELIVERY_KEY,
-                    {
-                        **build_delivery_record(pr, checks_summary, checks, branch, head_sha),
-                        "upstream": upstream or "",
-                        "checks_detail": checks_detail,
-                        "head_mismatch": False,
-                    },
-                ),
-            )
-        else:
-            update_workflow_state(
-                args.slug,
-                lambda state: state.__setitem__(DELIVERY_KEY, delivery_record),
-            )
-
-    else:
-        if args.dry_run:
-            print(f"branch: {branch}")
-            print(f"head: {head_sha}")
-            print("pr: not created (dry-run)")
-            return 0
-        update_workflow_state(
-            args.slug,
-            lambda state: state.__setitem__(
-                DELIVERY_KEY,
-                {
-                    "branch": branch,
-                    "head_sha": head_sha,
-                    "updated_at": int(time.time()),
-                    "upstream": upstream or "",
-                    "pr_url": "",
-                    "checks_summary": "not-started",
-                },
-            ),
-        )
-
-    delivery = load_workflow_state(args.slug).get(DELIVERY_KEY, {})
-    print(f"branch: {branch}")
-    print(f"head: {head_sha}")
-    if delivery.get("pr_url"):
-        print(f"pr: {delivery.get('pr_url')}")
-        print(f"checks: {delivery.get('checks_summary', 'unknown')}")
-        if delivery.get("merge_state_status"):
-            print(f"merge-state: {delivery.get('merge_state_status')}")
-        if delivery.get("head_mismatch"):
-            print("head-mismatch: yes")
-    else:
-        print("pr: not created")
-    _emit_next_action(args.slug, "deliver")
-    return 0
-
-
-def command_closeout(args: argparse.Namespace) -> int:
-    ensure_sync()
-    root = workflow_dir(args.slug)
-    reviews = gather_all_review_paths(args.slug, include_closeout=False)
-    plan_path = root / "plan.md"
-    state = load_workflow_state(args.slug)
-    delivery = refresh_delivery_snapshot(state.get(DELIVERY_KEY, {}))
-    if not delivery.get("pr_url"):
-        raise SystemExit("closeout requires a delivered PR first")
-    if delivery.get("checks_summary") != "pass":
-        raise SystemExit(
-            f"closeout requires passing checks first; "
-            f"current checks are {delivery.get('checks_summary', 'unknown')}"
-        )
-
-    missing_stages = []
-    for stage in VERIFY_ON_CLOSEOUT_STAGES:
-        manifest_path = checkpoint_manifest_path(args.slug, stage)
-        if not manifest_path.exists():
-            missing_stages.append(stage)
-    if missing_stages:
-        raise SystemExit(
-            "Closeout requires completed checkpoint manifests for: "
-            + ", ".join(missing_stages)
-        )
-
-    for stage in VERIFY_ON_CLOSEOUT_STAGES:
-        manifest_path = checkpoint_manifest_path(args.slug, stage)
-        run([sys.executable, str(VERIFY_CHECKPOINT), "--checkpoint-manifest", str(manifest_path)])
-
-    if reviews:
-        cmd = [
-            sys.executable,
-            str(VERIFY_RECONCILIATION),
-            "--plan",
-            str(plan_path),
-            "--require-terminal",
-        ]
-        for review in reviews:
-            cmd.extend(["--review", str(review)])
-        run(cmd)
-
-    summary_path = root / "closeout.md"
-    closeout_manifest_path = checkpoint_manifest_path(args.slug, "closeout")
-    dirty_override = state.get(DIRTY_OVERRIDE_KEY, {})
-    checkpoint_fallback = state.get(CHECKPOINT_FALLBACK_KEY, {})
-    backup_text = summary_path.read_text(encoding="utf-8") if summary_path.exists() else None
-    backup_manifest = closeout_manifest_path.read_text(encoding="utf-8") if closeout_manifest_path.exists() else None
-    inventory_text = closeout_inventory_text(args.slug, root, plan_path, reviews, delivery, dirty_override, checkpoint_fallback)
-    summary_text = compose_closeout_text(backup_text or "", inventory_text)
-    summary_path.write_text(summary_text, encoding="utf-8")
-
-    closeout_args = argparse.Namespace(
-        slug=args.slug,
-        stage="closeout",
-        artifact=str(summary_path),
-        base_ref=None,
-        context=[],
-        path=[],
-        extra_gemini_lens=[],
-        sensitive=False,
-        large_structural=False,
-        performance_sensitive=False,
-        use_existing_artifact=True,
-        allow_dirty_path=[],
-        max_artifact_chars=None,
-        max_context_chars=None,
-        max_total_chars=None,
-        gemini_timeout_seconds=120,
-        gemini_retries=1,
-        repair_timeout_seconds=300,
-        fallback=False,
-        allow_large_diff_rerun=False,
-    )
-    checkpoint_result = command_checkpoint(closeout_args)
-    if checkpoint_result != 0:
-        if backup_text is None:
-            summary_path.unlink(missing_ok=True)
-        else:
-            summary_path.write_text(backup_text, encoding="utf-8")
-        if backup_manifest is None:
-            closeout_manifest_path.unlink(missing_ok=True)
-        else:
-            closeout_manifest_path.write_text(backup_manifest, encoding="utf-8")
-        return checkpoint_result
-
-    refreshed_state = load_workflow_state(args.slug)
-    refreshed_delivery = refresh_delivery_snapshot(refreshed_state.get(DELIVERY_KEY, {}))
-    refreshed_inventory = closeout_inventory_text(
-        args.slug,
-        root,
-        plan_path,
-        reviews,
-        refreshed_delivery,
-        refreshed_state.get(DIRTY_OVERRIDE_KEY, {}),
-        refreshed_state.get(CHECKPOINT_FALLBACK_KEY, {}),
-    )
-    refreshed_summary = compose_closeout_text(backup_text or "", refreshed_inventory)
-    if refreshed_summary != summary_text:
-        summary_path.write_text(refreshed_summary, encoding="utf-8")
-        checkpoint_result = command_checkpoint(closeout_args)
-        if checkpoint_result != 0:
-            if backup_text is None:
-                summary_path.unlink(missing_ok=True)
-            else:
-                summary_path.write_text(backup_text, encoding="utf-8")
-            if backup_manifest is None:
-                closeout_manifest_path.unlink(missing_ok=True)
-            else:
-                closeout_manifest_path.write_text(backup_manifest, encoding="utf-8")
-            return checkpoint_result
-
-    print(str(summary_path))
-    _emit_next_action(args.slug, "closeout")
-    return 0
-
-
-def _codex_prompt_path(slug: str, i: int) -> Path:
-    safe_slug = "".join(ch if ch.isalnum() or ch in "-_" else "_" for ch in slug)
-    return Path(f"/tmp/codex-impl-{safe_slug}-{os.getpid()}-{i}.txt")
-
-
-def _build_codex_prompt(slug: str, i: int, tasks: list[str], file_scope: list[str]) -> str:
-    root = workflow_dir(slug)
-    # Prefer compact summaries — they contain everything Codex needs for implementation
-    # without the full narrative. Fall back to full files when summaries don't exist yet.
-    spec_path = root / "spec-acceptance.md" if (root / "spec-acceptance.md").exists() else root / "spec.md"
-    plan_path = root / "plan-summary.md" if (root / "plan-summary.md").exists() else root / "plan.md"
-
-    spec_content = spec_path.read_text(encoding="utf-8") if spec_path.exists() else ""
-    plan_content = plan_path.read_text(encoding="utf-8") if plan_path.exists() else ""
-
-    prompt_path = _codex_prompt_path(slug, i)
-    prompt_text = (
-        "# Implementation Task\n\n"
-        "## Context\n"
-        f"{spec_content}\n\n"
-        "## Plan\n"
-        f"{plan_content}\n\n"
-        "## Tasks to implement (your scope)\n"
-        f"{chr(10).join(map(str, tasks))}\n\n"
-        "## File scope\n"
-        f"{chr(10).join(map(str, file_scope)) if file_scope else '(no specific scope — implement all tasks)'}\n\n"
-        "Implement the tasks above. Commit your changes when done.\n"
-        "DO NOT MODIFY: CLAUDE.md, AGENTS.md, MEMORY.md, or any file not in your file scope.\n"
-    )
-    prompt_path.write_text(prompt_text, encoding="utf-8")
-    return prompt_text
-
-
-def _run_serial(slug: str, tasks: list[str]) -> int:
-    prompt_path = _codex_prompt_path(slug, 0)
-    prompt_text = _build_codex_prompt(slug, 0, tasks, [])
-    rc = 1
-    try:
-        result = subprocess.run(
-            ["codex", "exec", "-m", "gpt-5.4-mini", "-s", "workspace-write", prompt_text],
-            cwd=str(REPO_ROOT),
-            timeout=3600,
-        )
-        rc = result.returncode
-    except subprocess.TimeoutExpired:
-        print("[error] codex execution timed out after 3600 seconds", file=sys.stderr)
-        rc = 1
-    finally:
-        revert_protected_files()
-        prompt_path.unlink(missing_ok=True)
-    return rc
-
-
-def _run_parallel(slug: str, group: dict, max_workers: int = 4) -> int:
-    status = subprocess.run(
-        ["git", "-C", str(REPO_ROOT), "status", "--porcelain"],
-        capture_output=True,
-        text=True,
-    )
-    if status.returncode != 0:
-        print(
-            f"[error] unable to check working tree status: {status.stderr.strip() or status.stdout.strip() or 'git status failed'}",
-            file=sys.stderr,
-        )
-        return 1
-    if status.stdout.strip():
-        print("[error] working tree must be clean before implement", file=sys.stderr)
-        return 1
-
-    head_result = subprocess.run(
-        ["git", "-C", str(REPO_ROOT), "rev-parse", "HEAD"],
-        capture_output=True,
-        text=True,
-    )
-    if head_result.returncode != 0:
-        print(
-            f"[error] unable to capture base commit: {head_result.stderr.strip() or head_result.stdout.strip() or 'git rev-parse failed'}",
-            file=sys.stderr,
-        )
-        return 1
-    base_sha = head_result.stdout.strip()
-
-    worktree_paths: list[Path] = []
-    prompt_paths: list[Path] = []
-    failure_message = ""
-    failure = False
-    commits_by_task: dict[int, list[str]] = {}
-    executor = concurrent.futures.ThreadPoolExecutor(max_workers=max_workers)
-    try:
-        futures: dict[concurrent.futures.Future, int] = {}
-        tasks = list(group.get("tasks") or [])
-        file_scope = list(group.get("files") or [])
-        for i, task in enumerate(tasks):
-            try:
-                worktree_path = create_worktree(slug, i)
-                worktree_paths.append(worktree_path)
-                prompt_path = _codex_prompt_path(slug, i)
-                prompt_paths.append(prompt_path)
-                prompt_text = _build_codex_prompt(slug, i, [task], file_scope)
-                futures[
-                    executor.submit(
-                        subprocess.run,
-                        ["codex", "exec", "-m", "gpt-5.4-mini", "-s", "workspace-write", prompt_text],
-                        cwd=str(worktree_path),
-                        timeout=3600,
-                    )
-                ] = i
-            except Exception as exc:
-                failure = True
-                if not failure_message:
-                    failure_message = f"[error] failed to prepare codex worker {i}: {exc}"
-                break
-
-        try:
-            for future in concurrent.futures.as_completed(futures):
-                i = futures[future]
-                try:
-                    result = future.result()
-                    if result.returncode != 0 and not failure:
-                        failure = True
-                        failure_message = f"[error] codex worker {i} failed with return code {result.returncode}"
-                except Exception as exc:
-                    if not failure:
-                        failure = True
-                        failure_message = f"[error] codex worker {i} failed: {exc}"
-        finally:
-            executor.shutdown(wait=True, cancel_futures=True)
-
-        if failure:
-            reset_result = subprocess.run(
-                ["git", "-C", str(REPO_ROOT), "reset", "--hard", base_sha],
-                capture_output=True,
-                text=True,
-            )
-            if reset_result.returncode != 0:
-                print(
-                    f"[warn] failed to reset repository to {base_sha[:12]}: {reset_result.stderr.strip() or reset_result.stdout.strip() or 'git reset failed'}",
-                    file=sys.stderr,
-                )
-            print(failure_message, file=sys.stderr)
-            return 1
-
-        try:
-            for i in range(len(tasks)):
-                worktree_path = worktree_paths[i]
-                stage_and_commit_if_dirty(worktree_path, f"task {i}: auto-commit")
-                commits_by_task[i] = get_new_commits(worktree_path, base_sha)
-        except Exception as exc:
-            reset_result = subprocess.run(
-                ["git", "-C", str(REPO_ROOT), "reset", "--hard", base_sha],
-                capture_output=True,
-                text=True,
-            )
-            if reset_result.returncode != 0:
-                print(
-                    f"[warn] failed to reset repository to {base_sha[:12]}: {reset_result.stderr.strip() or reset_result.stdout.strip() or 'git reset failed'}",
-                    file=sys.stderr,
-                )
-            print(f"[error] failed to collect commits from worker worktrees: {exc}", file=sys.stderr)
-            return 1
-
-        all_commits = [c for i in sorted(commits_by_task) for c in commits_by_task[i]]
-        cherry_pick_ok, cherry_pick_detail = cherry_pick_commits(all_commits)
-        if not cherry_pick_ok:
-            reset_result = subprocess.run(
-                ["git", "-C", str(REPO_ROOT), "reset", "--hard", base_sha],
-                capture_output=True,
-                text=True,
-            )
-            if reset_result.returncode != 0:
-                print(
-                    f"[warn] failed to reset repository to {base_sha[:12]}: {reset_result.stderr.strip() or reset_result.stdout.strip() or 'git reset failed'}",
-                    file=sys.stderr,
-                )
-            print(f"[error] cherry-pick conflict: {cherry_pick_detail}", file=sys.stderr)
-            return 1
-
-        revert_protected_files()
-        return 0
-    finally:
-        remove_all_worktrees(worktree_paths)
-        for prompt_path in prompt_paths:
-            prompt_path.unlink(missing_ok=True)
-
-
-def command_implement(args: argparse.Namespace) -> int:
-    status = subprocess.run(
-        ["git", "-C", str(REPO_ROOT), "status", "--porcelain"],
-        capture_output=True,
-        text=True,
-    )
-    if status.returncode != 0:
-        print(
-            f"[error] unable to check working tree status: {status.stderr.strip() or status.stdout.strip() or 'git status failed'}",
-            file=sys.stderr,
-        )
-        return 1
-    if status.stdout.strip():
-        print("[error] working tree must be clean before implement", file=sys.stderr)
-        return 1
-
-    groups = parse_parallel_task_groups(args.slug)
-    if not groups:
-        print("nothing to implement — all tasks complete or no tasks.md")
-        return 0
-
-    for group in groups:
-        if not group["parallel"]:
-            if group.get("overlap_warning"):
-                print(f"[warn] {group['overlap_warning']} — running serially", file=sys.stderr)
-            rc = _run_serial(args.slug, group["tasks"])
-        else:
-            print(f"[implement] dispatching {len(group['tasks'])} parallel Codex workers...")
-            rc = _run_parallel(args.slug, group, max_workers=args.max_workers)
-        if rc != 0:
-            return rc
-    return 0
-
-
-def command_parallel(args: argparse.Namespace) -> int:
-    """Record whether the agent looked for parallel implementation opportunities.
-
-    Enforces that the agent explicitly considered parallelisation before the
-    tasks checkpoint. If --found is passed, validates that [P: file...] annotations
-    exist in tasks.md and that no two annotated tasks share a file (which would
-    cause a conflict at implement time).
-    """
-    ensure_sync()
-    note = (args.note or "").strip()
-    if not note:
-        raise SystemExit(
-            "parallel requires --note explaining what was found or why nothing "
-            "was safe to parallelise (e.g. 'all tasks share the schema migration')"
-        )
-    tasks_path = workflow_dir(args.slug) / "tasks.md"
-    if not tasks_path.exists():
-        raise SystemExit("parallel requires tasks.md to exist — write tasks first")
-
-    if args.found:
-        groups = parse_parallel_task_groups(args.slug)
-        parallel_groups = [g for g in groups if g["parallel"]]
-        if not parallel_groups:
-            raise SystemExit(
-                "parallel --found requires [P: file1, file2] annotations on tasks in "
-                "tasks.md — no valid parallel task groups detected. Add annotations or "
-                "omit --found if no safe parallelism exists."
-            )
-        for group in groups:
-            if group.get("overlap_warning"):
-                raise SystemExit(
-                    f"parallel --found blocked: {group['overlap_warning']} — "
-                    "parallel tasks must not share files. Fix the [P:] annotations "
-                    "before recording parallel opportunities."
-                )
-        print(f"[parallel] {len(parallel_groups)} parallel group(s) validated, no file conflicts")
-
-    def mutate(state: dict) -> None:
-        state[PARALLEL_ANALYSIS_KEY] = {
-            "reviewed": True,
-            "found": bool(args.found),
-            "note": note,
-            "updated_at": int(time.time()),
-        }
-
-    update_workflow_state(args.slug, mutate)
-    result = "opportunities found and validated" if args.found else "no safe opportunities found"
-    print(f"[parallel] analysis recorded: {result}")
-    print(f"[parallel] note: {note}")
-    _emit_next_action(args.slug, "parallel analysis")
     return 0
